fix: changed User API to singleton
