+++ src/app/dashboard/wagers/delete_wager/[id]/page.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { getOneWager, editWagerWithId } from "@/app/lib/getWagers";
+import { getOneWager, editWagerWithId } from "@/app/lib/data";
 import React, { useState, useEffect } from "react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
